Complete function getCargosOfCustomer

+++ src/main/java/org/apptopia/waterwayfreightsystem/api/api/core/model/Cargo.java
@@ -22,5 +22,5 @@ public class Cargo {
     private Integer idCargo;
 	@ManyToOne
 	private Account owner;
-	private Integer cargoWeight;
+	private Float cargoWeight;
 }
